src: fix clippy::module_inception for sink module
