+++ src/stream/sink/udp_sink.rs
@@ -4,7 +4,7 @@ use tracing::*;
 
 use gst::prelude::*;
 
-use super::sink::SinkInterface;
+use super::SinkInterface;
 use crate::stream::pipeline::PIPELINE_TEE_NAME;
 
 #[derive(Debug)]
